+++ gf180mcu/cells/cap_mos.py
@@ -2,8 +2,8 @@ import gdsfactory as gf
 import numpy as np
 from gdsfactory.typings import Float2, LayerSpec
 
+from gf180mcu.cells.via_generator import via_generator, via_stack
 from gf180mcu.layers import layer
-from gf180mcu.via_generator import via_generator, via_stack
 
 
 @gf.cell
