+++ gf180mcu/cells/diode.py
@@ -1,8 +1,8 @@
 import gdsfactory as gf
 from gdsfactory.typings import Float2
 
+from gf180mcu.cells.via_generator import via_generator, via_stack
 from gf180mcu.layers import layer
-from gf180mcu.via_generator import via_generator, via_stack
 
 
 @gf.cell
