+++ gf180mcu/cells/res.py
@@ -1,9 +1,9 @@
 import gdsfactory as gf
 from gdsfactory.typings import Float2, LayerSpec
 
-from gf180mcu.guardring import pcmpgr_gen
+from gf180mcu.cells.guardring import pcmpgr_gen
+from gf180mcu.cells.via_generator import via_stack
 from gf180mcu.layers import layer
-from gf180mcu.via_generator import via_stack
 
 
 @gf.cell
