+++ tests/test_components/test_pdk_settings_bend_s_.yml
@@ -1,0 +1,17 @@
+info:
+  end_angle: 0
+  length: 11.206
+  min_bend_radius: 13.012
+  route_info_length: 11.206
+  route_info_metal1_length: 11.206
+  route_info_min_bend_radius: 13.012
+  route_info_n_bend_s: 1
+  route_info_type: metal1
+  route_info_weight: 11.206
+  start_angle: 0
+name: bend_s_S11_1p8_CSmetal1_WNone
+settings:
+  cross_section: metal1
+  size:
+  - 11
+  - 1.8
