+++ tests/test_components/test_pdk_settings_wire_corner_.yml
@@ -1,0 +1,6 @@
+info:
+  dy: 0.2
+  length: 0.2
+name: wire_corner_CSmetal2_WNone
+settings:
+  cross_section: metal2
